Move CAS related stuff to api/cas/
