+++ api/cas/cache.py
@@ -6,7 +6,7 @@
 from api.cas.cli import *
 from api.cas.casadm_parser import *
 from test_utils.os_utils import *
-from cas_configuration.cache_config import *
+from api.cas.cache_config import *
 from storage_devices.device import Device
 from test_package.test_properties import TestProperties
 from api.cas.casadm_params import *
