+++ test_package/io_class/io_class_common.py
@@ -5,7 +5,7 @@
 
 from api.cas import casadm
 from api.cas import ioclass_config
-from cas_configuration.cache_config import CacheMode, CleaningPolicy
+from api.cas.cache_config import CacheMode, CleaningPolicy
 from storage_devices.disk import DiskType
 from test_package.conftest import base_prepare
 from test_package.test_properties import TestProperties
