+++ test_package/stats/test_block_stats.py
@@ -9,7 +9,7 @@ from api.cas.casadm import StatsFilter
 from api.cas import casadm
 from api.cas import ioclass_config
 from test_tools.dd import Dd
-from cas_configuration.cache_config import CacheMode, CleaningPolicy
+from api.cas.cache_config import CacheMode, CleaningPolicy
 from test_package.conftest import base_prepare
 from test_package.test_properties import TestProperties
 from storage_devices.disk import DiskType
